tests: data type fixes and typo fixes to make tests run

+++ apps/report-execution/tests/conftest.py
@@ -228,15 +228,16 @@ def insert_fake_data(
     with db_transaction(conn_string) as trx:
         # Tables with foreign keys pointing to the table we want to replace need to
         # be backed up and cleared out to avoid FK constraint violations
-        for fk_table in fk_tables:
-            temp_fk_table = temp_name(fk_table)
-            trx.execute(
-                f"IF OBJECT_ID('{temp_fk_table}') IS NOT NULL "
-                f'DROP TABLE {temp_fk_table}'
-            )
-            trx.execute(f'SELECT * INTO {temp_fk_table} FROM {fk_table}')
-            trx.execute(f'DELETE {fk_table}')
-            logging.info(f'cleared FK table: {fk_table}')
+        if fk_tables:
+            for fk_table in fk_tables:
+                temp_fk_table = temp_name(fk_table)
+                trx.execute(
+                    f"IF OBJECT_ID('{temp_fk_table}') IS NOT NULL "
+                    f'DROP TABLE {temp_fk_table}'
+                )
+                trx.execute(f'SELECT * INTO {temp_fk_table} FROM {fk_table}')
+                trx.execute(f'DELETE {fk_table}')
+                logging.info(f'cleared FK table: {fk_table}')
 
         for db_table in db_tables:
             tmp_db_table = temp_name(db_table)
